+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Breaking changes:
 - [plugin] 'Hosted mode' extracted in `plugin-dev` extension
 - [core] `scheme` is mandatory for URI
   - `URI.withoutScheme` is removed, in order to get a path use `URI.path`
+- [plugin-ext] improved QuickInput and InputBox API's
 
 ## v0.7.0